Removed redundant Storefront interface and renamed ItemStorefront to Storefront.
